// src/api/generated/models/nutritionAlertsResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * ECD Backend API
 * Early Childhood Development management system API.
 * OpenAPI spec version: 1.0
 */
import type { NutritionAlertDto } from './nutritionAlertDto';

export interface NutritionAlertsResponseDto {
  items: NutritionAlertDto[];
  total: number;
}